# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/permissions": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get Permissions",
                "description": "Get a list of permissions",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of results for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Permission"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Create Permission",
                "description": "Create Permission by name, description",
                "operationId": "store",
                "requestBody": {
                    "description": "Send name, value",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "create-users"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Create Users"
                                    },
                                    "create": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "update": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "read": {
                                        "type": "string",
                                        "example": "on"
                                    },
                                    "delete": {
                                        "type": "string",
                                        "example": "on"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/permissions/{id}": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get Permission",
                "description": "Show Permission by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permission"
                ],
                "summary": "Update Permission",
                "description": "Update Permission",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Send name or display_name or description to update permission",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "create-users"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Create Users"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Create Users"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "Delete Permission",
                "description": "Delete Permission by id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get Role",
                "description": "Show Role by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolePermissions"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Role"
                ],
                "summary": "Update Role",
                "description": "Update Role",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Send name, value to update role",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "student"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "Student"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Student"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Delete Role",
                "description": "Delete role by id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "description": "Show user data by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRoles"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "Update User",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Send name, email, photo to update user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user@email.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "description": "Delete user by id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Permission": {
                "required": [
                    "name",
                    "display_name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "3"
                    },
                    "name": {
                        "description": "Name of the role",
                        "type": "string",
                        "example": "create-users"
                    },
                    "display_name": {
                        "description": "Display name of the role",
                        "type": "string",
                        "example": "Create Users"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Create Users"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Permission"
                }
            },
            "Role": {
                "required": [
                    "name",
                    "display_name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "3"
                    },
                    "name": {
                        "description": "Name of the role",
                        "type": "string",
                        "example": "student"
                    },
                    "display_name": {
                        "description": "Display name of the role",
                        "type": "string",
                        "example": "Student"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Student"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Role"
                }
            },
            "RolePermissions": {
                "required": [
                    "name",
                    "display_name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "3"
                    },
                    "name": {
                        "description": "Name of the role",
                        "type": "string",
                        "example": "student"
                    },
                    "display_name": {
                        "description": "Display name of the role",
                        "type": "string",
                        "example": "Student"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Student"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Role"
                }
            },
...
            "UserRoles": {
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@email.com"
                    },
                    "photo": {
                        "description": "User photo",
                        "type": "string",
                        "format": "string",
                        "example": "/images/johndoe.png"
                    },
                    "created_at": {
                        "description": "Datetime of when user was created",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "roles": {
                        "$ref": "#/components/schemas/Role"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserRoles"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer {access_token}",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}